+Lombok, small refactorings

+++ ConnectedVehicles-Customer/src/main/java/com/youssef/connectedvehicles/repository/CustomerRepository.java
@@ -8,6 +8,6 @@ import org.springframework.stereotype.Repository;
 import com.youssef.connectedvehicles.entity.Customer;
 
 @Repository
-public interface CustomerRepository extends MongoRepository<Customer, Long>{
+public interface CustomerRepository extends MongoRepository<Customer, Integer>{
 	List<Customer> findByName(String name);
 }
